tests/wpt/tests/encoding-detection/fa-windows-1256.tentative.html

<!doctype html>
<title>fa windows-1256</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>«&#1740;‰ &#1740;ò  ”  —„“ê–«—&#1740; ò«—«ò — «” .</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1256", 'Expected windows-1256');
    }, "Check detection result");
      done();
};
</script>
